requirements.txt
drf_spectacular==0.27.2
Django==5.0.3
djangorestframework==3.15.1